+++ src/main/java/io/github/thebusybiscuit/slimefun4/implementation/items/electric/machines/FluidPump.java
@@ -4,6 +4,7 @@ import java.util.List;
 import java.util.Optional;
 
 import org.bukkit.Material;
+import org.bukkit.World;
 import org.bukkit.block.Block;
 import org.bukkit.block.BlockFace;
 import org.bukkit.block.data.BlockData;
@@ -34,7 +35,7 @@ import me.mrCookieSlime.Slimefun.api.inventory.BlockMenu;
 import me.mrCookieSlime.Slimefun.api.inventory.BlockMenuPreset;
 
 /**
- * This machine draws liquids from the world and puts them
+ * This machine collects liquids from the {@link World} and puts them
  * into buckets provided to the machine by using energy.
  *
  * @author TheBusyBiscuit
@@ -140,34 +141,47 @@ public class FluidPump extends SimpleSlimefunItem<BlockTicker> implements Invent
     }
 
     private void consumeFluid(Block fluid) {
-        if (!fluid.isLiquid()) return;
-        if (fluid.getType() == Material.WATER) {
-            fluid.setType(Material.AIR);
-            return;
+        if (fluid.isLiquid()) {
+            if (fluid.getType() == Material.WATER || fluid.getType() == Material.BUBBLE_COLUMN) {
+                // With water we can be sure to find an infinite source whenever we go
+                // further than 2 blocks, so we can just remove the water here and save
+                // outselves all of that computing...
+                fluid.setType(Material.AIR);
+            }
+            else {
+                List<Block> list = Vein.find(fluid, RANGE, block -> isLiquid(block) && block.getType() == fluid.getType());
+                list.get(list.size() - 1).setType(Material.AIR);
+            }
         }
-
-        List<Block> list = Vein.find(fluid, RANGE, block -> isLiquid(block) && block.getType() == fluid.getType());
-        list.get(list.size() - 1).setType(Material.AIR);
     }
 
     private Optional<ItemStack> getFilledBucket(Block fluid) {
         if (fluid.getType() == Material.LAVA) {
             return Optional.of(new ItemStack(Material.LAVA_BUCKET));
         }
-        else if (fluid.getType() == Material.WATER) {
+        else if (fluid.getType() == Material.WATER || fluid.getType() == Material.BUBBLE_COLUMN) {
             return Optional.of(new ItemStack(Material.WATER_BUCKET));
         }
-
-        return Optional.empty();
-    }
-    
-    private boolean isLiquid(Block block) {
-        if (!block.isLiquid()) return false;
-        BlockData data = block.getBlockData();
-        if (data instanceof Levelled) {
-            return ((Levelled) data).getLevel() == 0;
+        else {
+            return Optional.empty();
         }
-        return false;
+    }
+
+    private boolean isLiquid(Block block) {
+        if (block.isLiquid()) {
+            BlockData data = block.getBlockData();
+
+            if (data instanceof Levelled) {
+                // Check if this is a full block.
+                return ((Levelled) data).getLevel() == 0;
+            }
+            else {
+                return false;
+            }
+        }
+        else {
+            return false;
+        }
     }
 
     @Override