+++ CHANGELOG.md
@@ -83,6 +83,8 @@
 * Fixed #2243
 * Fixed #2249
 * Fixed #1022
+* Fixed Fluid Pump treating low-level fluids like stationary fluids
+* Fixed Fluid Pump not working on Bubble Columns
 * Fixed #2208
 
 ## Release Candidate 15 (01 Aug 2020)